Update snapshots, reorganize code
